refactor: extract haptic priming logic into useHapticPriming hook
